Migrate cssVarsPonyfill
